main.js 파일분리
aside, list 관련 함수 파일 분리
