+++ TeamProject/static/js/board/event.js
@@ -2,6 +2,7 @@ import * as MAIN from "./main.js"
 import * as REND from "./render.js"
 import * as FETCH from "./fetch.js"
 import * as REND_LIST from "./list/render.js";
+import * as LIST from "./list/index.js"
 //===========보드 메인 포스트 페이지 ==========
 
 //메인화면 페이지로 가는 함수
@@ -186,7 +187,7 @@ export const handle_scrollHeight = async () => {
       const ele = document.querySelector('.post_loading');
       ele.parentNode.removeChild(ele);
       const hashValue = location.hash.split('#');
-      MAIN.loading_new_post(hashValue);
+      LIST.loading_new_post(hashValue);
     }, 500);
     setTimeout(() => {
       console.log('1초뒤');
@@ -221,7 +222,7 @@ export function handle_drop() { //drag&drop
     const data = event.dataTransfer;
     const MAX_FILE = 5;
     drop_zone.style.cssText = "border: 3px dashed lightgray;";
-    INPUT_DATA_FILE.append_file(data.files); //파일 객체에 추가
+    MAIN.INPUT_DATA_FILE.append_file(data.files); //파일 객체에 추가
   });
 }
