+++ TeamProject/templates/post.html
@@ -90,6 +90,8 @@
 <script type="module" src="/static/js/board/router.js"></script>
 <script type="module" src="/static/js/board/list/render.js"></script>
 <script type="module" src="/static/js/board/aside/render.js"></script>
+<script type="module" src="/static/js/board/list/index.js"></script>
+<script type="module" src="/static/js/board/aside/main.js"></script>
 
 <script type="module" src="/static/js/auth.js"></script>
